Rewrite shell decomposition on tree-sitter-bash AST
Replace the hand-rolled string walker with a grammar-driven AST visitor
built on tree-sitter-bash. The trigger was the bash `'\''`
close-escape-open idiom inside a `$(...)` substitution misclassifying as
unknown — symptomatic of a class of quote-state bugs that the string
walker couldn't avoid without reimplementing chunks of the bash grammar
piecemeal. Closes #4.

Architecture:

  - hooks/grammar_classifier.py — tree-sitter-bash AST visitor; the public
    entry point. Walks `command`, `pipeline`, `list`, `redirected_statement`,
    `for_statement`, `if_statement`, `while_statement`, `case_statement`,
    `subshell`, `compound_statement`, `command_substitution`,
    `process_substitution`, `variable_assignment`, `function_definition`,
    and so on, dispatching per node kind. Argv reconstructed from typed
    argument children (`word`, `string`, `raw_string`, `concatenation`,
    `simple_expansion`, ...).
  - hooks/rule_classifier.py — argv-level rule lookup against
    rules/shell.json. Extracted from the previous ShellClassifier with
    decomposition removed; its inputs are tokens, not raw bash strings.
  - hooks/yolt_analyzer.py — Python AST analyzer (unchanged in substance);
    the hook entry now imports GrammarClassifier and exits silently if
    tree-sitter / tree-sitter-bash aren't importable.
  - hooks/shell_classifier.py — deleted.

Behavioral wins over the old walker:

  - Handles `'\''` idiom inside `$(...)` correctly (the trigger bug).
  - `<(rm -rf /a)` inside a process substitution now surfaces as unsafe
    instead of being treated as opaque.
  - Heredocs, multi-line backslash continuations, comments, ANSI-C `$'...'`
    strings, arithmetic `$((...))`, and parameter expansion `${...//x/y}`
    handled by the grammar rather than as edge cases.
  - `function_definition` no longer classified as if its body were running.

Deps: tree-sitter, tree-sitter-bash (both pure Python via wheels).
Documented in requirements.txt and README. Hook stays silent on import
error so a broken install doesn't break the user's session.

Tests: 132 passing. New file split mirrors the architecture
(test_rule_classifier for argv helpers, test_grammar_classifier for
end-to-end via classify(), test_yolt_hook for the subprocess hook).
