+++ hooks/rule_classifier.py
@@ -1,0 +1,493 @@
+#!/usr/bin/env python3
+"""YOLT argv-level rule classifier.
+
+Given a list of argv tokens for a single simple command, classify the call
+as safe / unsafe / unknown by consulting `rules/shell.json`. This module is
+language-agnostic w.r.t. Bash quoting and decomposition; it expects the
+caller (the grammar-driven walker in `grammar_classifier.py`) to have
+already produced argv from the AST.
+
+The rule schema lives in `rules/shell.json`. Major shapes:
+
+  - `commands`: dict of cmd_name -> spec. Spec defaults: "safe", "unsafe",
+    "ask", "subcommand", "delegate_to_argument", "aws_cli", "gcloud_cli",
+    "az_cli". Subcommand specs carry safe / unsafe lists and patterns.
+  - `interpreters`: dict of cmd_name -> spec for `python3 -c`, `bash -c`,
+    `python3 file.py`. Inline-flag scripts get delegated to a Python AST
+    analyzer; bash inline scripts re-enter via the grammar walker (the
+    grammar walker passes itself in as the bash analyzer).
+  - `shell_builtins_safe`: list of always-safe builtins.
+
+Settings I/O (`load_shell_rules`, `load_allow_patterns`, `match_allow_patterns`)
+also lives here for symmetry.
+"""
+
+import json
+import os
+import re
+from fnmatch import fnmatch
+from pathlib import Path
+
+
+DECISION_SAFE = "safe"
+DECISION_UNSAFE = "unsafe"
+DECISION_UNKNOWN = "unknown"
+
+SUBSTITUTION_PLACEHOLDER = "__YOLT_SUB__"
+
+ASSIGNMENT_RE = re.compile(r"^[A-Za-z_][A-Za-z0-9_]*=")
+BASH_PATTERN_RE = re.compile(r"^Bash\((.*)\)$")
+
+
+def load_shell_rules(rules_dir, user_overrides_path=None):
+    """Load shell rules from rules_dir/shell.json plus optional overrides.
+    Overrides merge per top-level key, one level deep."""
+    rules = {}
+
+    default_path = Path(rules_dir) / "shell.json"
+    if default_path.exists():
+        with open(default_path, "r") as f:
+            rules = json.load(f)
+
+    if user_overrides_path and Path(user_overrides_path).exists():
+        with open(user_overrides_path, "r") as f:
+            overrides = json.load(f)
+        for key, value in overrides.items():
+            if key in rules and isinstance(rules[key], dict) and isinstance(value, dict):
+                rules[key].update(value)
+            else:
+                rules[key] = value
+
+    return rules
+
+
+def load_allow_patterns(settings_paths):
+    """Read Claude Code settings.json files and return the list of inner
+    Bash() allow patterns (the part between the parentheses)."""
+    patterns = []
+    seen = set()
+    for path in settings_paths:
+        if not path:
+            continue
+        p = Path(path)
+        if not p.exists():
+            continue
+        try:
+            with open(p, "r") as f:
+                data = json.load(f)
+        except (OSError, ValueError):
+            continue
+        permissions = data.get("permissions") or {}
+        allow = permissions.get("allow") or []
+        if not isinstance(allow, list):
+            continue
+        for entry in allow:
+            if not isinstance(entry, str):
+                continue
+            m = BASH_PATTERN_RE.match(entry.strip())
+            if not m:
+                continue
+            inner = m.group(1).strip()
+            if not inner or inner in seen:
+                continue
+            seen.add(inner)
+            patterns.append(inner)
+    return patterns
+
+
+def match_allow_patterns(command, patterns):
+    """Return the matching pattern if `command` matches any allow pattern,
+    else None. Matching is fnmatch against the unwrapped inner pattern."""
+    if not patterns:
+        return None
+    cmd = command.strip()
+    for pat in patterns:
+        if fnmatch(cmd, pat):
+            return pat
+    return None
+
+
+def aggregate_decisions(decisions):
+    """Combine (decision, reason) results. Precedence: unsafe > unknown > safe."""
+    if not decisions:
+        return (DECISION_SAFE, "nothing to classify")
+
+    unsafe_reasons = [r for d, r in decisions if d == DECISION_UNSAFE]
+    unknown_reasons = [r for d, r in decisions if d == DECISION_UNKNOWN]
+    safe_reasons = [r for d, r in decisions if d == DECISION_SAFE]
+
+    if unsafe_reasons:
+        return (DECISION_UNSAFE, "; ".join(unsafe_reasons))
+    if unknown_reasons:
+        return (DECISION_UNKNOWN, "; ".join(unknown_reasons))
+    return (DECISION_SAFE, "; ".join(safe_reasons) if safe_reasons else "no commands")
+
+
+def check_unsafe_flags(cmd_args, spec):
+    """Check whether cmd_args contains a flag combination the command spec
+    declares unsafe. Returns a human-readable description, or None."""
+    unsafe_flag_values = spec.get("unsafe_flag_values", {})
+    unsafe_flag_any_value = set(spec.get("unsafe_flag_any_value", []))
+    unsafe_flags_without_value = set(spec.get("unsafe_flags_without_value", []))
+
+    i = 0
+    while i < len(cmd_args):
+        tok = cmd_args[i]
+
+        flag = tok
+        inline_value = None
+        if "=" in tok and tok.startswith("-"):
+            flag, _, inline_value = tok.partition("=")
+
+        if flag in unsafe_flags_without_value:
+            return "{} (no value)".format(flag)
+
+        if flag in unsafe_flag_values:
+            value = inline_value
+            if value is None and i + 1 < len(cmd_args):
+                value = cmd_args[i + 1]
+            if value is not None:
+                value_upper = value.upper() if isinstance(value, str) else value
+                for unsafe_val in unsafe_flag_values[flag]:
+                    if value_upper == unsafe_val.upper():
+                        return "{} {}".format(flag, value)
+
+        if flag in unsafe_flag_any_value:
+            return flag
+
+        i += 1
+
+    return None
+
+
+def parse_aws_positionals(cmd_args):
+    """Walk aws CLI args skipping flags. Return (service, operation, trailing)."""
+    service = None
+    operation = None
+    trailing = []
+
+    i = 0
+    while i < len(cmd_args):
+        tok = cmd_args[i]
+        if tok.startswith("--"):
+            if "=" in tok:
+                i += 1
+                continue
+            if tok in {"--no-cli-pager", "--no-paginate", "--no-verify-ssl",
+                      "--no-sign-request", "--debug"}:
+                i += 1
+                continue
+            if i + 1 < len(cmd_args) and not cmd_args[i + 1].startswith("-"):
+                i += 2
+                continue
+            i += 1
+            continue
+        if tok.startswith("-") and len(tok) > 1:
+            i += 1
+            continue
+        if service is None:
+            service = tok
+        elif operation is None:
+            operation = tok
+        else:
+            trailing.append(tok)
+        i += 1
+
+    return (service, operation, trailing)
+
+
+class RuleClassifier:
+    """Classifies a simple command (argv tokens already reconstructed) by
+    consulting rule data. The grammar-driven walker owns decomposition;
+    this class is purely the rule-lookup half."""
+
+    MAX_RECURSION_DEPTH = 8
+
+    def __init__(self, rules, python_analyzer_factory=None, bash_analyzer=None):
+        self.rules = rules
+        self.commands = rules.get("commands", {})
+        self.shell_builtins_safe = set(rules.get("shell_builtins_safe", []))
+        self.interpreters = rules.get("interpreters", {})
+        self.python_analyzer_factory = python_analyzer_factory
+        # When a `bash -c '<script>'` interpreter call needs nested analysis,
+        # the grammar walker injects itself here as a callable
+        # (script: str) -> (decision, reason).
+        self.bash_analyzer = bash_analyzer
+
+    def classify_tokens(self, tokens, _depth=0):
+        if _depth > self.MAX_RECURSION_DEPTH:
+            return (DECISION_UNKNOWN, "max recursion depth")
+
+        if not tokens:
+            return (DECISION_SAFE, "empty")
+
+        if all(t == SUBSTITUTION_PLACEHOLDER for t in tokens):
+            return (DECISION_SAFE, "only substitution result (classified separately)")
+
+        cmd_name = os.path.basename(tokens[0])
+        cmd_args = tokens[1:]
+
+        if cmd_name in self.shell_builtins_safe:
+            return (DECISION_SAFE, "builtin: {}".format(cmd_name))
+
+        if cmd_name in self.interpreters:
+            return self.classify_interpreter(cmd_name, cmd_args, _depth=_depth)
+
+        if cmd_name in self.commands:
+            return self.classify_known_command(cmd_name, cmd_args, _depth=_depth)
+
+        return (DECISION_UNKNOWN, "no rule: {}".format(cmd_name))
+
+    def classify_interpreter(self, name, cmd_args, _depth):
+        spec = self.interpreters[name]
+        inline_flag = spec.get("inline_flag")
+        delegate = spec.get("delegate", "unknown")
+        read_script_file = spec.get("read_script_file", False)
+
+        inline_code = None
+        if inline_flag and inline_flag in cmd_args:
+            idx = cmd_args.index(inline_flag)
+            if idx + 1 < len(cmd_args):
+                inline_code = cmd_args[idx + 1]
+
+        if inline_code is not None:
+            if delegate == "python":
+                return self.classify_python_source(
+                    inline_code, "{} {} ...".format(name, inline_flag)
+                )
+            if delegate == "bash":
+                if self.bash_analyzer is None:
+                    return (DECISION_UNKNOWN, "bash analyzer unavailable")
+                d, r = self.bash_analyzer(inline_code, _depth + 1)
+                return (d, "{} -c -> {}".format(name, r))
+            return (DECISION_UNKNOWN, "cannot analyze inline {}".format(delegate))
+
+        if read_script_file and delegate == "python":
+            for arg in cmd_args:
+                if arg.startswith("-"):
+                    continue
+                if arg.endswith(".py") or os.path.isfile(arg):
+                    try:
+                        with open(arg, "r") as f:
+                            code = f.read()
+                    except (OSError, UnicodeDecodeError) as e:
+                        return (DECISION_UNKNOWN, "cannot read {}: {}".format(arg, e))
+                    return self.classify_python_source(code, "{} {}".format(name, arg))
+                break
+
+        return (DECISION_UNKNOWN, "{} invocation not analyzable".format(name))
+
+    def classify_python_source(self, source, description):
+        if self.python_analyzer_factory is None:
+            return (DECISION_UNKNOWN, "python analyzer unavailable")
+        analyzer = self.python_analyzer_factory()
+        result = analyzer.analyze(source)
+        if result.get("safe"):
+            return (DECISION_SAFE, "python: {}".format(description))
+        return (DECISION_UNSAFE, "python {}: {}".format(
+            description, result.get("reason", "destructive call")
+        ))
+
+    def classify_known_command(self, cmd_name, cmd_args, _depth):
+        spec = self.commands[cmd_name]
+        default = spec.get("default", "unknown")
+
+        unsafe_match = check_unsafe_flags(cmd_args, spec)
+        if unsafe_match:
+            return (DECISION_UNSAFE, "{}: flag {}".format(cmd_name, unsafe_match))
+
+        if default == "safe":
+            return (DECISION_SAFE, "{}: read-only".format(cmd_name))
+        if default == "unsafe":
+            return (DECISION_UNSAFE, "{}: mutating".format(cmd_name))
+        if default == "ask":
+            return (DECISION_UNKNOWN, "{}: rules punt".format(cmd_name))
+        if default == "subcommand":
+            return self.classify_subcommand(cmd_name, cmd_args, spec, _depth=_depth)
+        if default == "delegate_to_argument":
+            return self.classify_delegate_to_argument(cmd_name, cmd_args, spec, _depth=_depth)
+        if default == "aws_cli":
+            return self.classify_aws(cmd_args, spec)
+        if default == "gcloud_cli":
+            return self.classify_verb_cli(cmd_args, spec, label="gcloud")
+        if default == "az_cli":
+            return self.classify_verb_cli(cmd_args, spec, label="az")
+
+        return (DECISION_UNKNOWN, "{}: unknown default '{}'".format(cmd_name, default))
+
+    def classify_subcommand(self, cmd_name, cmd_args, spec, _depth):
+        valueless = set(spec.get("valueless_flags", []))
+        positional = self._skip_flags(cmd_args, valueless_flags=valueless)
+        if not positional:
+            return (DECISION_UNKNOWN, "{}: no subcommand".format(cmd_name))
+
+        sub = positional[0]
+        rest = positional[1:]
+        remaining_args = cmd_args[cmd_args.index(sub) + 1:] if sub in cmd_args else rest
+
+        nested = spec.get("nested_subcommand", {})
+        if sub in nested:
+            nested_spec = nested[sub]
+            nested_unsafe = check_unsafe_flags(remaining_args, nested_spec)
+            if nested_unsafe:
+                return (DECISION_UNSAFE, "{} {}: flag {}".format(cmd_name, sub, nested_unsafe))
+            if "default" in nested_spec:
+                nested_default = nested_spec["default"]
+                if nested_default == "safe":
+                    return (DECISION_SAFE, "{} {}: read-only".format(cmd_name, sub))
+                if nested_default == "unsafe":
+                    return (DECISION_UNSAFE, "{} {}: mutating".format(cmd_name, sub))
+            return self._match_subcommand_lists(
+                "{} {}".format(cmd_name, sub), rest, nested_spec
+            )
+
+        return self._match_subcommand_lists(cmd_name, positional, spec)
+
+    def _match_subcommand_lists(self, label, positional, spec):
+        safe_subs = set(spec.get("safe_subcommands", []))
+        unsafe_subs = set(spec.get("unsafe_subcommands", []))
+        safe_patterns = spec.get("safe_subcommand_patterns", [])
+        unsafe_patterns = spec.get("unsafe_subcommand_patterns", [])
+
+        if not positional:
+            return (DECISION_UNKNOWN, "{}: no subcommand".format(label))
+
+        sub = positional[0]
+
+        if sub in safe_subs:
+            return (DECISION_SAFE, "{} {}: read-only".format(label, sub))
+        if sub in unsafe_subs:
+            return (DECISION_UNSAFE, "{} {}: mutating".format(label, sub))
+        for pat in safe_patterns:
+            if fnmatch(sub, pat):
+                return (DECISION_SAFE, "{} {}: matches safe pattern {}".format(label, sub, pat))
+        for pat in unsafe_patterns:
+            if fnmatch(sub, pat):
+                return (DECISION_UNSAFE, "{} {}: matches unsafe pattern {}".format(label, sub, pat))
+
+        return (DECISION_UNKNOWN, "{} {}: no rule".format(label, sub))
+
+    def classify_delegate_to_argument(self, cmd_name, cmd_args, spec, _depth):
+        skip_first_positional = spec.get("_skip_first_positional", False)
+
+        i = 0
+        while i < len(cmd_args):
+            tok = cmd_args[i]
+            if tok.startswith("--"):
+                if "=" in tok:
+                    i += 1
+                    continue
+                if i + 1 < len(cmd_args) and not cmd_args[i + 1].startswith("-"):
+                    i += 2
+                    continue
+                i += 1
+                continue
+            if tok.startswith("-") and len(tok) > 1:
+                i += 1
+                continue
+            if cmd_name == "env" and ASSIGNMENT_RE.match(tok) and "=" in tok:
+                i += 1
+                continue
+            if skip_first_positional:
+                skip_first_positional = False
+                i += 1
+                continue
+            break
+
+        if i >= len(cmd_args):
+            return (DECISION_UNKNOWN, "{}: no wrapped command".format(cmd_name))
+
+        wrapped_tokens = cmd_args[i:]
+        d, r = self.classify_tokens(wrapped_tokens, _depth=_depth + 1)
+        return (d, "{} wraps: {}".format(cmd_name, r))
+
+    def classify_aws(self, cmd_args, spec):
+        service, operation, _ = parse_aws_positionals(cmd_args)
+        if service is None:
+            return (DECISION_UNKNOWN, "aws: no service")
+        if operation is None:
+            return (DECISION_UNKNOWN, "aws {}: no operation".format(service))
+
+        service_overrides = spec.get("service_overrides", {})
+        per_service = service_overrides.get(service, {})
+
+        safe_subs = set(per_service.get("safe_subcommands", []))
+        unsafe_subs = set(per_service.get("unsafe_subcommands", []))
+        if operation in safe_subs:
+            return (DECISION_SAFE, "aws {} {}: read-only".format(service, operation))
+        if operation in unsafe_subs:
+            return (DECISION_UNSAFE, "aws {} {}: mutating".format(service, operation))
+
+        for pat in per_service.get("extra_safe_patterns", []):
+            if fnmatch(operation, pat):
+                return (DECISION_SAFE, "aws {} {}: service override safe".format(service, operation))
+
+        for pat in per_service.get("safe_operation_patterns", []):
+            if fnmatch(operation, pat):
+                return (DECISION_SAFE, "aws {} {}: service-safe pattern".format(service, operation))
+        for pat in per_service.get("unsafe_operation_patterns", []):
+            if fnmatch(operation, pat):
+                return (DECISION_UNSAFE, "aws {} {}: service-unsafe pattern".format(service, operation))
+
+        for pat in spec.get("safe_operation_patterns", []):
+            if fnmatch(operation, pat):
+                return (DECISION_SAFE, "aws {} {}: global read-only".format(service, operation))
+        for pat in spec.get("unsafe_operation_patterns", []):
+            if fnmatch(operation, pat):
+                return (DECISION_UNSAFE, "aws {} {}: global mutating".format(service, operation))
+
+        return (DECISION_UNKNOWN, "aws {} {}: unclassified".format(service, operation))
+
+    def classify_verb_cli(self, cmd_args, spec, label):
+        positional = self._skip_flags(cmd_args)
+        if len(positional) < 2:
+            return (DECISION_UNKNOWN, "{}: too few positionals".format(label))
+
+        verb = positional[-1]
+        safe_verbs = {"list", "describe", "get", "show", "export", "version", "help"}
+        unsafe_verbs = {
+            "create", "delete", "update", "replace", "set", "add", "remove",
+            "deploy", "run", "execute", "exec", "ssh",
+            "enable", "disable", "reset", "rollback", "promote", "restart",
+            "scale", "resize",
+        }
+        if verb in safe_verbs:
+            return (DECISION_SAFE, "{} {}: read verb".format(label, verb))
+        if verb in unsafe_verbs:
+            return (DECISION_UNSAFE, "{} {}: write verb".format(label, verb))
+        return (DECISION_UNKNOWN, "{} {}: unknown verb".format(label, verb))
+
+    @staticmethod
+    def _skip_flags(cmd_args, valueless_flags=None):
+        valueless_flags = valueless_flags or set()
+        positional = []
+        i = 0
+        while i < len(cmd_args):
+            tok = cmd_args[i]
+            if tok.startswith("--"):
+                if "=" in tok:
+                    i += 1
+                    continue
+                if tok in valueless_flags:
+                    i += 1
+                    continue
+                if i + 1 < len(cmd_args) and not cmd_args[i + 1].startswith("-"):
+                    i += 2
+                    continue
+                i += 1
+                continue
+            if tok.startswith("-") and len(tok) > 1:
+                if tok in valueless_flags:
+                    i += 1
+                    continue
+                if len(tok) > 2 and not tok[2].isalpha():
+                    i += 1
+                    continue
+                if i + 1 < len(cmd_args) and not cmd_args[i + 1].startswith("-"):
+                    i += 2
+                    continue
+                i += 1
+                continue
+            positional.append(tok)
+            i += 1
+        return positional
